Move geo utils to /lib/geo
