+++ src/init/consented/prepare-a9.ts
@@ -3,9 +3,9 @@ import { once } from 'lodash-es';
 import { a9Apstag } from '../../core/__vendor/a9-apstag';
 import { commercialFeatures } from '../../lib/commercial-features';
 import { isGoogleProxy } from '../../lib/detect/detect-google-proxy';
+import { isInCanada } from '../../lib/geo/geo-utils';
 import { a9 } from '../../lib/header-bidding/a9/a9';
 import { shouldIncludeOnlyA9 } from '../../lib/header-bidding/utils';
-import { isInCanada } from '../../utils/geo-utils';
 
 const shouldLoadA9 = () =>
 	// There are two articles that InfoSec would like to avoid loading scripts on
