+++ src/lib/header-bidding/slot-config.ts
@@ -1,7 +1,7 @@
 import type { AdSize } from '../../core/ad-sizes';
 import { adSizes, createAdSize } from '../../core/ad-sizes';
 import type { Advert } from '../../define/Advert';
-import { isInUk } from '../../utils/geo-utils';
+import { isInUk } from '../../lib/geo/geo-utils';
 import type {
 	HeaderBiddingSizeKey,
 	HeaderBiddingSizeMapping,
